pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datahub-cicd"
version = "0.1.0"
description = "CI/CD pipeline for DataHub governance metadata"
requires-python = ">=3.10"
dependencies = [
    "acryl-datahub>=1.0.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
